tasks/Cheryl/requirements.txt
Flask
flask-login
Flask-SqlAlchemy